source_files/requirements.txt
pyperclip
PILLOW
psutil